Improved FPS disconnect prevention
The original patch still caused you to go into offline mode when you returned to the main menu if your FPS dropped in-game. Changing a conditional jump to a mandatory jump upstream of where the original patch is applied means that the value that is checked when you return to the main menu is changed and you can carry on in online mode without restarting the game.